Fifth blog added.
Seems to be an error where the image is not added if the text is too big.
The error can be avoided by adding the title and some small text in the description when creating a post. Then add the remaining description by editing the post after it has been created.